+++ src/Service/Generation/vpg_object_factory_file_generation_service.cpp
@@ -1,4 +1,4 @@
-#include "vpg_property_accessor_factory_file_generation_service.hpp"
+#include "vpg_object_factory_file_generation_service.hpp"
 
 #include <set>
 #include <string>
@@ -10,12 +10,12 @@
 
 using namespace vcc;
 
-#define LOG_ID L"Prorperty Accessor Factory File Generation"
+#define LOG_ID L"Object Factory File Generation"
 
-void VPGPropertyAccessorFactoryFileGenerationSerive::GenerateHpp(const LogProperty *logProperty, const std::wstring &hppFilePath)
+void VPGObjectFactoryFileGenerationSerive::GenerateHpp(const LogProperty *logProperty, const std::wstring &hppFilePath)
 {
     TRY
-        LogService::LogInfo(logProperty, LOG_ID, L"Generate property accessor factory file: " + hppFilePath);
+        LogService::LogInfo(logProperty, LOG_ID, L"Generate object factory file: " + hppFilePath);
         std::wstring content = L""
             "#pragma once\r\n"
             "\r\n"
@@ -23,37 +23,36 @@ void VPGPropertyAccessorFactoryFileGenerationSerive::GenerateHpp(const LogProper
             "\r\n"
             "#include \"base_factory.hpp\"\r\n"
             "#include \"i_object.hpp\"\r\n"
-            "#include \"i_property_accessor.hpp\"\r\n"
+            "#include \"object_type.hpp\"\r\n"
             "\r\n"
             "using namespace vcc;\r\n"
             "\r\n"
-            "class PropertyAccessorFactory : public BaseFactory\r\n"
+            "class ObjectFactory : public BaseFactory\r\n"
             "{\r\n"
             "    private:\r\n"
-            "        PropertyAccessorFactory() = default;\r\n"
-            "        virtual ~PropertyAccessorFactory() {}\r\n"
+            "        ObjectFactory() = default;\r\n"
+            "        virtual ~ObjectFactory() {}\r\n"
             "\r\n"
             "    public:\r\n"
-            "        static std::shared_ptr<IPropertyAccessor> Create(std::shared_ptr<IObject> object);\r\n"
+            "        static std::shared_ptr<IObject> Create(const ObjectType &objectType);\r\n"
             "};\r\n";
         WriteFile(hppFilePath, content, true);
-        LogService::LogInfo(logProperty, LOG_ID, L"Generate property accessor factory file completed.");
+        LogService::LogInfo(logProperty, LOG_ID, L"Generate object factory file completed.");
     CATCH
 }
 
-void VPGPropertyAccessorFactoryFileGenerationSerive::GenerateCpp(const LogProperty *logProperty, const std::wstring &projectPrefix, const std::set<std::wstring> &includeFiles,
+void VPGObjectFactoryFileGenerationSerive::GenerateCpp(const LogProperty *logProperty, const std::wstring &projectPrefix, const std::set<std::wstring> &includeFiles,
     const std::wstring &cppFilePath, const std::set<std::wstring> &propertyTypes)
 {
     TRY
-        LogService::LogInfo(logProperty, LOG_ID, L"Generate property accessor factory file: " + cppFilePath);
+        LogService::LogInfo(logProperty, LOG_ID, L"Generate object factory file: " + cppFilePath);
 
         std::set<std::wstring> tmpIncludePaths = includeFiles;
-        tmpIncludePaths.insert(L"base_property_accessor.hpp");
         tmpIncludePaths.insert(L"i_object.hpp");
-        tmpIncludePaths.insert(L"i_property_accessor.hpp");
+        tmpIncludePaths.insert(L"object_type.hpp");
 
         std::wstring content = L""
-            "#include \"property_accessor_factory.hpp\"\r\n"
+            "#include \"object_factory.hpp\"\r\n"
             "\r\n"
             "#include <assert.h>\r\n"
             "#include <memory>\r\n"
@@ -65,15 +64,13 @@ void VPGPropertyAccessorFactoryFileGenerationSerive::GenerateCpp(const LogProper
         content += L"\r\n"
             "using namespace vcc;\r\n"
             "\r\n"
-            "std::shared_ptr<IPropertyAccessor> PropertyAccessorFactory::Create(std::shared_ptr<IObject> object)\r\n"
+            "std::shared_ptr<IObject> ObjectFactory::Create(const ObjectType &objectType)\r\n"
             "{\r\n"
-            + INDENT + L"assert(object != nullptr);\r\n"
-            "\r\n"
-            + INDENT + L"switch (object->GetObjectType())\r\n"
+            + INDENT + L"switch (objectType)\r\n"
             + INDENT + L"{\r\n";
         for (auto const &propertyType : propertyTypes) {
             content += INDENT + L"case ObjectType::" + propertyType + L":\r\n"
-                + INDENT + INDENT + L"return std::make_shared<" + projectPrefix + propertyType + L"PropertyAccessor>(object);\r\n";
+                + INDENT + INDENT + L"return std::make_shared<" + projectPrefix + propertyType + L">();\r\n";
         }
         content += L""
             + INDENT + L"default:\r\n"
@@ -83,6 +80,6 @@ void VPGPropertyAccessorFactoryFileGenerationSerive::GenerateCpp(const LogProper
             + INDENT + L"return nullptr;\r\n"
             "}\r\n";
         WriteFile(cppFilePath, content, true);
-        LogService::LogInfo(logProperty, LOG_ID, L"Generate property accessor factory completed.");
+        LogService::LogInfo(logProperty, LOG_ID, L"Generate object factory completed.");
     CATCH
 }