+++ unittest/Service/Generation/vpg_object_factory_file_generation_service_test.cpp
@@ -7,14 +7,14 @@
 #include "class_macro.hpp"
 #include "file_helper.hpp"
 
-#include "vpg_property_accessor_factory_file_generation_service.hpp"
+#include "vpg_object_factory_file_generation_service.hpp"
 
 using namespace vcc;
 
-class VPGPropertyAccessorFactoryFileGenerationServiceTest : public testing::Test 
+class VPGObjectFactoryFileGenerationServiceTest : public testing::Test 
 {
     GETSET_SPTR(LogProperty, LogProperty);
-    GETSET(std::wstring, Workspace, L"bin/Debug/VPGPropertyAccessorFactoryFileGenerationServiceTest/");
+    GETSET(std::wstring, Workspace, L"bin/Debug/VPGObjectFactoryFileGenerationServiceTest/");
     
     GETSET(std::wstring, FilePathHpp, L"");
     GETSET(std::wstring, FilePathCpp, L"");
@@ -27,8 +27,8 @@ class VPGPropertyAccessorFactoryFileGenerationServiceTest : public testing::Test
             this->_LogProperty->SetIsConsoleLog(false);
             std::filesystem::remove_all(PATH(this->GetWorkspace()));
 
-            this->_FilePathHpp = ConcatPaths({this->GetWorkspace(), L"property_accessor_factory.hpp"});
-            this->_FilePathCpp = ConcatPaths({this->GetWorkspace(), L"property_accessor_factory.cpp"});
+            this->_FilePathHpp = ConcatPaths({this->GetWorkspace(), L"object_factory.hpp"});
+            this->_FilePathCpp = ConcatPaths({this->GetWorkspace(), L"object_factory.cpp"});
 
             this->_ExpectedHpp = L""
                 "#pragma once\r\n"
@@ -37,18 +37,18 @@ class VPGPropertyAccessorFactoryFileGenerationServiceTest : public testing::Test
                 "\r\n"
                 "#include \"base_factory.hpp\"\r\n"
                 "#include \"i_object.hpp\"\r\n"
-                "#include \"i_property_accessor.hpp\"\r\n"
+                "#include \"object_type.hpp\"\r\n"
                 "\r\n"
                 "using namespace vcc;\r\n"
                 "\r\n"
-                "class PropertyAccessorFactory : public BaseFactory\r\n"
+                "class ObjectFactory : public BaseFactory\r\n"
                 "{\r\n"
                 "    private:\r\n"
-                "        PropertyAccessorFactory() = default;\r\n"
-                "        virtual ~PropertyAccessorFactory() {}\r\n"
+                "        ObjectFactory() = default;\r\n"
+                "        virtual ~ObjectFactory() {}\r\n"
                 "\r\n"
                 "    public:\r\n"
-                "        static std::shared_ptr<IPropertyAccessor> Create(std::shared_ptr<IObject> object);\r\n"
+                "        static std::shared_ptr<IObject> Create(const ObjectType &objectType);\r\n"
                 "};\r\n";
         }
 
@@ -58,11 +58,11 @@ class VPGPropertyAccessorFactoryFileGenerationServiceTest : public testing::Test
         }
 };
 
-TEST_F(VPGPropertyAccessorFactoryFileGenerationServiceTest, Empty)
+TEST_F(VPGObjectFactoryFileGenerationServiceTest, Empty)
 {
     std::set<std::wstring> propertyTypes;
-    VPGPropertyAccessorFactoryFileGenerationSerive::GenerateHpp(this->GetLogProperty().get(), this->GetFilePathHpp());
-    VPGPropertyAccessorFactoryFileGenerationSerive::GenerateCpp(this->GetLogProperty().get(), L"VCC", {}, this->GetFilePathCpp(), propertyTypes);
+    VPGObjectFactoryFileGenerationSerive::GenerateHpp(this->GetLogProperty().get(), this->GetFilePathHpp());
+    VPGObjectFactoryFileGenerationSerive::GenerateCpp(this->GetLogProperty().get(), L"VCC", {}, this->GetFilePathCpp(), propertyTypes);
 
     EXPECT_TRUE(IsFileExists(this->GetFilePathHpp()));
     EXPECT_TRUE(IsFileExists(this->GetFilePathCpp()));
@@ -70,22 +70,19 @@ TEST_F(VPGPropertyAccessorFactoryFileGenerationServiceTest, Empty)
     EXPECT_EQ(ReadFile(this->GetFilePathHpp()), this->GetExpectedHpp());
 
     std::wstring expectedResult = L""
-        "#include \"property_accessor_factory.hpp\"\r\n"
+        "#include \"object_factory.hpp\"\r\n"
         "\r\n"
         "#include <assert.h>\r\n"
         "#include <memory>\r\n"
         "\r\n"
-        "#include \"base_property_accessor.hpp\"\r\n"
         "#include \"i_object.hpp\"\r\n"
-        "#include \"i_property_accessor.hpp\"\r\n"
+        "#include \"object_type.hpp\"\r\n"
         "\r\n"
         "using namespace vcc;\r\n"
         "\r\n"
-        "std::shared_ptr<IPropertyAccessor> PropertyAccessorFactory::Create(std::shared_ptr<IObject> object)\r\n"
+        "std::shared_ptr<IObject> ObjectFactory::Create(const ObjectType &objectType)\r\n"
         "{\r\n"
-        "    assert(object != nullptr);\r\n"
-        "\r\n"
-        "    switch (object->GetObjectType())\r\n"
+        "    switch (objectType)\r\n"
         "    {\r\n"
         "    default:\r\n"
         "        assert(false);\r\n"
@@ -96,13 +93,13 @@ TEST_F(VPGPropertyAccessorFactoryFileGenerationServiceTest, Empty)
     EXPECT_EQ(ReadFile(this->GetFilePathCpp()), expectedResult);
 }
 
-TEST_F(VPGPropertyAccessorFactoryFileGenerationServiceTest, Normal)
+TEST_F(VPGObjectFactoryFileGenerationServiceTest, Normal)
 {
     std::set<std::wstring> propertyTypes;
     propertyTypes.insert(L"Def");
     propertyTypes.insert(L"Abc");
-    VPGPropertyAccessorFactoryFileGenerationSerive::GenerateHpp(this->GetLogProperty().get(), this->GetFilePathHpp());
-    VPGPropertyAccessorFactoryFileGenerationSerive::GenerateCpp(this->GetLogProperty().get(), L"VCC", { L"abc.hpp" }, this->GetFilePathCpp(), propertyTypes);
+    VPGObjectFactoryFileGenerationSerive::GenerateHpp(this->GetLogProperty().get(), this->GetFilePathHpp());
+    VPGObjectFactoryFileGenerationSerive::GenerateCpp(this->GetLogProperty().get(), L"VCC", { L"abc.hpp" }, this->GetFilePathCpp(), propertyTypes);
 
     EXPECT_TRUE(IsFileExists(this->GetFilePathHpp()));
     EXPECT_TRUE(IsFileExists(this->GetFilePathCpp()));
@@ -110,28 +107,25 @@ TEST_F(VPGPropertyAccessorFactoryFileGenerationServiceTest, Normal)
     EXPECT_EQ(ReadFile(this->GetFilePathHpp()), this->GetExpectedHpp());
 
     std::wstring expectedResult = L""
-        "#include \"property_accessor_factory.hpp\"\r\n"
+        "#include \"object_factory.hpp\"\r\n"
         "\r\n"
         "#include <assert.h>\r\n"
         "#include <memory>\r\n"
         "\r\n"
         "#include \"abc.hpp\"\r\n"
-        "#include \"base_property_accessor.hpp\"\r\n"
         "#include \"i_object.hpp\"\r\n"
-        "#include \"i_property_accessor.hpp\"\r\n"
+        "#include \"object_type.hpp\"\r\n"
         "\r\n"
         "using namespace vcc;\r\n"
         "\r\n"
-        "std::shared_ptr<IPropertyAccessor> PropertyAccessorFactory::Create(std::shared_ptr<IObject> object)\r\n"
+        "std::shared_ptr<IObject> ObjectFactory::Create(const ObjectType &objectType)\r\n"
         "{\r\n"
-        "    assert(object != nullptr);\r\n"
-        "\r\n"
-        "    switch (object->GetObjectType())\r\n"
+        "    switch (objectType)\r\n"
         "    {\r\n"
         "    case ObjectType::Abc:\r\n"
-        "        return std::make_shared<VCCAbcPropertyAccessor>(object);\r\n"
+        "        return std::make_shared<VCCAbc>();\r\n"
         "    case ObjectType::Def:\r\n"
-        "        return std::make_shared<VCCDefPropertyAccessor>(object);\r\n"
+        "        return std::make_shared<VCCDef>();\r\n"
         "    default:\r\n"
         "        assert(false);\r\n"
         "        break;\r\n"