+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 - Makefile MacOS extension changed from .so to .dylib
+- vcc.json rename ModelDirectory to ObjectDirectory
+- Object Factory, Property Accessor Factory
 
 ### [v0.0.5] - 2024-06-02: Basic Strucute: Property Accessor and Thread safe
 - Generate Mode supports generate Property Accessor ReadWrite, ReadOnly, WriteOnly, NoAccess